feat(portal): Quests カードから問題詳細ページに遷移可能に
#163 残作業の一部。/problems/:problemId placeholder を実機能に置換し、
Quests のカード headers をクリッカブルにした。

## 変更点

- 新規 `components/ProblemPanel.tsx`: Home の `ProblemCard` + `FlagSubmissionPanel`
  を抽出した共通 component (Home / ProblemDetail 両方から使う)
- 新規 `pages/ProblemDetail.tsx`: useTeamView() の問題リストから :problemId 一致
  の 1 件を抽出し ProblemPanel を表示。問題不在 (deploy 無し / 旧 deployment) は
  Alert で説明
- `pages/Home.tsx`: 抽出元のコードを ProblemPanel import に置換 (= shrink)
- `pages/Quests.tsx`: カード header に Cloudscape `Link` を入れて
  `/problems/:problemId` に navigate
- `App.tsx`: /problems/:problemId route を Placeholder → ProblemDetailPage に置換

## 物理影響

| 種別 | リソース | 注 |
| --- | --- | --- |
| UPDATE | participant-portal SPA | 新ページ + shared ProblemPanel + route swap |
| NO-OP  | Lambda / IAM / DDB / CDK | 不変 |

## Test plan

- [x] make before-commit 緑 (425 件全 pass、UI 重構成のみ unit test 不変)
- [ ] AWS で実 deploy → Quests のカード header クリック → 詳細ページ navigation
- [ ] 詳細ページから「問題一覧へ戻る」で Quests に戻れる
- [ ] flag 提出 / uptime 表示は Home と同等 (= ProblemPanel 共有)

## 残 TODO (本 PR 範囲外)

- /problems/:problemId に問題説明 (description / learningGoals from metadata.json)
  の表示 — 現状は team view の data のみ。metadata 取得 API は追加検討
- Battle / Challenge UI の分離 (Issue #502) は別 PR
